models/requirements.txt
# Unified requirements.txt
# Main application dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0

# Machine learning / Deep learning dependencies
torch>=2.0.0
torchvision>=0.15.0
scikit-learn

# Notes:
# - torch, torchvision maintain minimum version requirements (>=2.0.0, >=0.15.0)
# - scikit-learn added for classifier usage
# - Includes all packages from original main requirements.txt
